finish C++11 port; add more tests to testing suite

+++ test/Arduino.h
@@ -6,7 +6,8 @@
 typedef uint8_t byte;
 
 // Minimal Arduino API used by the library. Tests provide the implementations.
-unsigned long millis();
-void delay(unsigned long ms);
+// millis() returns uint32_t to match the 32-bit timer on all Arduino targets.
+uint32_t millis();
+void delay(uint32_t ms);
 
 #endif // ARDUINO_SHIM_H
